fix(docx): cast addCell widths to int (PHPWord @param int $width, Converter returns float)
PHPWord's Row::addCell() and Table::addCell() declare $width as int
(@param int). The Converter helpers that produce twip values
(inchToTwip, pixelToTwip) are @return float because the multiplication
(INCH_TO_TWIP, / INCH_TO_PIXEL * INCH_TO_TWIP) is floating-point. At
three call sites in the DOCX content renderer the float was fed
straight into addCell, which the IDE flags as PHP0406 ("Argument '1'
passed to addCell() is expected to be of type int|null, float given").

Twip units are integral by spec; PHPWord itself performs integer
arithmetic on the value once stored. Cast the three widths with (int)
so the value narrows to int at the API boundary and the IDE diagnostic
disappears. No behavior change at runtime.

+++ includes/class-sscribe-docx-content-renderer.php
@@ -721,7 +721,7 @@ class SScribe_DOCX_Content_Renderer {
 					$effective_width = $cell_width * $colspan;
 				}
 
-				$cell_obj = $table->addCell( $effective_width, $cell_style );
+				$cell_obj = $table->addCell( (int) $effective_width, $cell_style );
 				if ( ! empty( $cell['runs'] ) ) {
 					$text_run = $cell_obj->addTextRun( $this->get_para_style() );
 					$this->render_runs( $text_run, $cell['runs'], false, ! empty( $cell['is_header'] ) );
@@ -755,7 +755,7 @@ class SScribe_DOCX_Content_Renderer {
 		);
 
 		$table->addRow();
-		$cell = $table->addCell( Converter::inchToTwip( 5 ), array( 'bgColor' => $this->colors['light_bg'] ) );
+		$cell = $table->addCell( (int) Converter::inchToTwip( 5 ), array( 'bgColor' => $this->colors['light_bg'] ) );
 
 		$cell->addText(
 			'[ACTION BUTTON] ' . $this->safe_text( ! empty( $element['content'] ) ? $element['content'] : __( 'Click Here', 'sscribe-export-site-pages' ) ),
@@ -956,7 +956,7 @@ class SScribe_DOCX_Content_Renderer {
 				)
 			);
 			$table->addRow();
-			$cell = $table->addCell( Converter::inchToTwip( 5.5 ), array( 'bgColor' => 'F8FAFC' ) );
+			$cell = $table->addCell( (int) Converter::inchToTwip( 5.5 ), array( 'bgColor' => 'F8FAFC' ) );
 			$cell->addText(
 				__( 'IMAGE ASSET SOURCE URL:', 'sscribe-export-site-pages' ),
 				array(
